DROTH-3515_3516 move database operations with IncompleteLinks to RoadLinkOverrideDAO
